chore: update footer
